Start separating entitiesData from repository
- so that we can start loading it from a file
- for #29
